featue: 新增资源筛选器API逻辑；pipeline_plugins 模块代码优化 #1724
