+++ pipeline_plugins/components/query/sites/open/query.py
@@ -22,8 +22,7 @@ from django.conf.urls import url
 from auth_backend.constants import AUTH_FORBIDDEN_CODE
 from auth_backend.exceptions import AuthFailedException
 
-from pipeline_plugins.components.utils import (
-    cc_get_inner_ip_by_module_id,
+from pipeline_plugins.base.utils.inject import (
     supplier_account_inject,
     supplier_id_inject
 )
@@ -142,34 +141,16 @@ def cc_format_topo_data(data, obj_id, category):
         if category == "prev":
             if item['bk_obj_id'] != obj_id:
                 tree_data.append(tree_item)
-                if item.get('child'):
+                if 'child' in item:
                     tree_item['children'] = cc_format_topo_data(item['child'], obj_id, category)
         else:
             if item['bk_obj_id'] == obj_id:
                 tree_data.append(tree_item)
-            elif item.get('child'):
+            elif 'child' in item:
                 tree_item['children'] = cc_format_topo_data(item['child'], obj_id, category)
                 tree_data.append(tree_item)
 
     return tree_data
-
-
-def cc_format_module_hosts(username, biz_cc_id, module_id_list, supplier_account):
-    module_host_list = cc_get_inner_ip_by_module_id(username, biz_cc_id, module_id_list, supplier_account)
-    module_host_dict = {}
-    for item in module_host_list:
-        for module in item['module']:
-            if module_host_dict.get('module_%s' % module['bk_module_id']):
-                module_host_dict['module_%s' % module['bk_module_id']].append({
-                    'id': '%s_%s' % (module['bk_module_id'], item['host']['bk_host_innerip']),
-                    'label': item['host']['bk_host_innerip']
-                })
-            else:
-                module_host_dict['module_%s' % module['bk_module_id']] = [{
-                    'id': '%s_%s' % (module['bk_module_id'], item['host']['bk_host_innerip']),
-                    'label': item['host']['bk_host_innerip']
-                }]
-    return module_host_dict
 
 
 @supplier_account_inject
@@ -196,31 +177,12 @@ def cc_search_topo(request, obj_id, category, biz_cc_id, supplier_account):
         }
         return JsonResponse(result)
 
-    if category in ["normal", "prev", "picker"]:
+    if category in ["normal", "prev"]:
         cc_topo = cc_format_topo_data(cc_result['data'], obj_id, category)
     else:
         cc_topo = []
 
     return JsonResponse({'result': True, 'data': cc_topo})
-
-
-@supplier_account_inject
-def cc_get_host_by_module_id(request, biz_cc_id, supplier_account):
-    """
-    查询模块对应主机
-    :param request:
-    :param biz_cc_id:
-    :return:
-    """
-    select_module_id = request.GET.getlist('query', [])
-    # 查询module对应的主机
-    module_hosts = cc_format_module_hosts(request.user.username, biz_cc_id, [int(x) for x in select_module_id],
-                                          supplier_account)
-
-    for del_id in (set(module_hosts.keys()) - set(['module_%s' % x for x in select_module_id])):
-        del module_hosts[del_id]
-
-    return JsonResponse({'result': True, 'data': module_hosts})
 
 
 def job_get_script_list(request, biz_cc_id):
@@ -519,7 +481,6 @@ urlpatterns = [
     url(r'^cc_search_object_attribute/(?P<obj_id>\w+)/(?P<biz_cc_id>\d+)/$', cc_search_object_attribute),
     url(r'^cc_search_create_object_attribute/(?P<obj_id>\w+)/(?P<biz_cc_id>\d+)/$', cc_search_create_object_attribute),
     url(r'^cc_search_topo/(?P<obj_id>\w+)/(?P<category>\w+)/(?P<biz_cc_id>\d+)/$', cc_search_topo),
-    url(r'^cc_get_host_by_module_id/(?P<biz_cc_id>\d+)/$', cc_get_host_by_module_id),
     url(r'^job_get_script_list/(?P<biz_cc_id>\d+)/$', job_get_script_list),
     url(r'^job_get_own_db_account_list/(?P<biz_cc_id>\d+)/$', job_get_own_db_account_list),
     url(r'^file_upload/(?P<project_id>\d+)/$', file_upload),
